backend-new/requirements.txt
fastapi
uvicorn
python-multipart
bcrypt
google-cloud-firestore
google-cloud-datastore
google-auth